requirements.txt
flask
flask-cors
python-dotenv
azure-storage-blob
requests
